feat: update book endpoints

+++ backend/Routes/Books.cs
@@ -39,11 +39,14 @@ public static class BookEndpoints
         }).RequireAuthorization();
 
         // GET hämtar en specifik bok
-        app.MapGet("/book/{id}", async (AppDbContext db, int id) =>
+        app.MapGet("/book/{id}", async (AppDbContext db, ClaimsPrincipal user, int id) =>
         {
             try
             {
-                var book = await db.Books.FindAsync(id);
+                var userId = PasswordService.GetUserId(user);
+                if (userId == null) return Results.Unauthorized();
+
+                var book = await db.Books.FirstOrDefaultAsync(b => b.UserId == userId && b.UserId == userId);
                 if (book is null) return Results.NotFound();
 
                 return Results.Ok(book);
@@ -52,14 +55,19 @@ public static class BookEndpoints
             {
                 return Results.Problem(ex.Message);
             }
-        });
+        }).RequireAuthorization();
 
         // POST skapa en bok
         // Mappar det inkommande JSON objektet till ett Book objekt genom Book book
-        app.MapPost("/book", async (AppDbContext db, Book book) =>
+        app.MapPost("/book", async (AppDbContext db, ClaimsPrincipal user, Book book) =>
         {
             try
             {
+                var userId = PasswordService.GetUserId(user);
+                if (userId == null) return Results.Unauthorized();
+
+                //Sätter user ID från tokenen
+                book.UserId = userId.Value;
                 db.Books.Add(book);
                 await db.SaveChangesAsync();
                 return Results.Created($"/{book.Id}", book); //201
@@ -68,19 +76,24 @@ public static class BookEndpoints
             {
                 return Results.Problem(ex.Message);
             }
-        });
+        }).RequireAuthorization();
 
         // PUT uppdaterar en bok
-        app.MapPut("/book/{id}", async (AppDbContext db, int id, Book book) =>
+        app.MapPut("/book/{id}", async (AppDbContext db, ClaimsPrincipal user, int id, Book book) =>
         {
             try
             {
+                var userId = PasswordService.GetUserId(user);
+                if (userId == null) return Results.Unauthorized();
+
                 // Om det inte finns en bok med det angivna id:t
                 // AnyAsync returnerar ett boolean värde medan FindAsync returnerar ett helt objekt
-                if (!await db.Books.AnyAsync(e => e.Id == id)) return Results.NotFound();
+                if (!await db.Books.AnyAsync(e => e.Id == id && e.UserId == userId)) return Results.NotFound();
 
                 //Sätter bokens Id från URL:en så att EF Core vet vilken bok som ska uppdateras. Kommer annars skapa en ny resurs
                 book.Id = id;
+                //säkerställer att userId inte ändras
+                book.UserId = userId.Value;
 
                 db.Books.Update(book);
                 await db.SaveChangesAsync();
@@ -91,15 +104,17 @@ public static class BookEndpoints
             {
                 return Results.Problem(ex.Message);
             }
-        });
+        }).RequireAuthorization();
 
         // DELETE ta bort en bok
-        app.MapDelete("/book/{id}", async (AppDbContext db, int id) =>
+        app.MapDelete("/book/{id}", async (AppDbContext db, ClaimsPrincipal user, int id) =>
         {
             try
             {
-                var book = await db.Books.FindAsync(id);
+                var userId = PasswordService.GetUserId(user);
+                if (userId == null) return Results.Unauthorized();
 
+                var book = await db.Books.FirstOrDefaultAsync(b => b.Id == id && b.UserId == userId);
                 if (book is null) return Results.NotFound();
 
                 db.Books.Remove(book);
@@ -110,6 +125,6 @@ public static class BookEndpoints
             {
                 return Results.Problem(ex.Message);
             }
-        });
+        }).RequireAuthorization();
     }
 }
